feat: add user_id on home, search, records router

+++ routes/modules/records.js
@@ -9,10 +9,9 @@ router.get('/new', (req, res) => {
 
 // create a new record
 router.post('/new', async (req, res, next) => {
-  const { name, date, category, amount } = req.body
-  const user_id = req.user._id
-
   try {
+    const user_id = req.user._id
+    const { name, date, category, amount } = req.body
     const categoryData = await Category.findOne({ name: category }).lean()
     await Record.create({
       name,
@@ -29,9 +28,10 @@ router.post('/new', async (req, res, next) => {
 
 //  get edit record page
 router.get('/:id/edit', async (req, res, next) => {
-  const _id = req.params.id
   try {
-    const record = await Record.findOne({ _id }).lean().populate('category_id').lean()
+    const user_id = req.user._id
+    const _id = req.params.id
+    const record = await Record.findOne({ _id, user_id }).lean().populate('category_id').lean()
     record.date = record.date.toISOString().slice(0, 10)  //  轉字串並取前10位
     res.render('edit', { record })
   } catch (e) {
@@ -42,9 +42,10 @@ router.get('/:id/edit', async (req, res, next) => {
 //  edit a record
 router.put('/:id', async (req, res, next) => {
   try {
+    const user_id = req.user._id
     const _id = req.params.id
     const { name, date, category, amount } = req.body
-    const categoryData = await Category.findOne({ name: category }).lean()
+    const categoryData = await Category.findOne({ name: category, user_id }).lean()
     await Record.findOneAndUpdate({ _id }, { name, date, category_id: categoryData._id, amount })
     res.redirect('/')
   } catch (e) {
@@ -55,8 +56,9 @@ router.put('/:id', async (req, res, next) => {
 //  delete a record
 router.delete('/:id', async (req, res, next) => {
   try {
+    const user_id = req.user._id
     const _id = req.params.id
-    await Record.findByIdAndDelete({ _id })
+    await Record.findByIdAndDelete({ _id, user_id })
     res.redirect('/')
   } catch (e) {
     next(e)
